requirements.txt
numpy~=1.25.2
pandas~=2.1.1
openpyxl~=3.1.2
bs4~=0.0.1
beautifulsoup4~=4.12.2
requests~=2.20.0
aiohttp~=3.8.6
colorama~=0.4.6
censys~=2.2.7
click~=8.0.1
Levenshtein~=0.23.0
jieba~=0.42.1
nltk~=3.8.1
pwn~=1.0
shodan~=1.30.1
scikit-learn~=1.3.2
matplotlib~=3.7.3
pyecharts~=2.0.4
mysql.connector